% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{check_package}
\alias{check_package}
\title{Check that required packages are available (for Suggests dependencies)}
\usage{
check_package(pkg)
}
\description{
Check that required packages are available (for Suggests dependencies)
}
\keyword{internal}
