% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{e2splits}
\alias{e2splits}
\alias{e2splits.e2tree}
\title{Extract Split Information from an E2Tree Model}
\usage{
e2splits(x, ...)

\method{e2splits}{e2tree}(x, ...)
}
\arguments{
\item{x}{An e2tree object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with components:
\describe{
\item{splits}{The split information matrix.}
\item{csplit}{The categorical split encoding matrix.}
}
}
\description{
Returns the split matrix and categorical split encoding from a fitted
E2Tree model.
}
