% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2tree-package.R
\docType{package}
\name{e2tree-package}
\alias{e2tree}
\alias{e2tree-package}
\title{e2tree: Creating An Explainable Tree From An Ensemble Tree Structure}
\description{
e2tree implements a methodological proposal to provide an explanable tree structure for ensemble tree methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/massimoaria/e2tree}
  \item Report bugs at \url{https://github.com/massimoaria/e2tree/issues}
}

}
\author{
\strong{Maintainer}: Massimo Aria \email{aria@unina.it} (\href{https://orcid.org/0000-0002-8517-9411}{ORCID}) [copyright holder]

}
\keyword{internal}
