% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2tree.R
\name{e2tree}
\alias{e2tree}
\title{Explainable Ensemble Tree}
\usage{
e2tree(
  D,
  X,
  response,
  setting = list(impTotal = 0.1, maxDec = 0.01, n = 5, level = 5, tMax = 5)
)
}
\arguments{
\item{D}{is the dissimilarity matrix}

\item{X}{is the training data frame with only the predictors}

\item{response}{is the vector of responses of the training dataset}

\item{setting}{is a list containing the setting parameters for tree building procedure.
Default is \code{setting=list(impTotal=0.1, maxDec=0.01, n=5, level=5, tMax=5)}.}
}
\value{
a e2tree object.

#examples
}
\description{
It creates an explainable tree for Random Forest
}
