% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNeighbors.R
\name{plot.e2neighbors}
\alias{plot.e2neighbors}
\title{Plot method for e2neighbors}
\usage{
\method{plot}{e2neighbors}(x, obs = 1L, ...)
}
\arguments{
\item{x}{An \code{e2neighbors} object.}

\item{obs}{Integer. Index of the query observation to plot. Default \code{1}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Horizontal bar chart of the nearest ensemble neighbours of one query
observation, with bars ordered by proximity and (for classification)
coloured by the neighbour's outcome.
}
