% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_e2tree.R
\name{plot.e2tree}
\alias{plot.e2tree}
\title{Plot an E2Tree model}
\usage{
\method{plot}{e2tree}(x, ensemble = NULL, main = "E2Tree", ...)
}
\arguments{
\item{x}{An e2tree object}

\item{ensemble}{The ensemble model (randomForest or ranger).
Required for converting the tree to rpart format.}

\item{main}{Plot title. Default is "E2Tree".}

\item{...}{Additional arguments passed to \code{rpart.plot::rpart.plot}}
}
\description{
Displays the tree structure using rpart.plot.
This is a convenience wrapper around \code{\link{plot_e2tree}}.
}
