% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goi.R
\name{plot.goi_perm}
\alias{plot.goi_perm}
\title{Plot method for Permutation Test results}
\usage{
\method{plot}{goi_perm}(x, ...)
}
\arguments{
\item{x}{A goi_perm object}

\item{...}{Additional arguments passed to hist()}
}
\description{
Displays the null distribution with the observed statistic and
confidence intervals.
}
