% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_e2tree.R
\name{print.e2tree}
\alias{print.e2tree}
\title{Print an E2Tree model}
\usage{
\method{print}{e2tree}(x, ...)
}
\arguments{
\item{x}{An e2tree object}

\item{...}{Additional arguments (ignored)}
}
\description{
Displays a compact summary of the fitted E2Tree model including
task type, tree size, terminal nodes, and splitting variables.
}
