% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e2tree_vis.R
\name{save_e2tree_html}
\alias{save_e2tree_html}
\title{Save E2Tree visNetwork Plot to HTML}
\usage{
save_e2tree_html(vis, file = "e2tree_plot.html", selfcontained = TRUE)
}
\arguments{
\item{vis}{A visNetwork object from plot_e2tree_vis()}

\item{file}{Output file path (should end with .html)}

\item{selfcontained}{Include all dependencies in single file}
}
\description{
Save E2Tree visNetwork Plot to HTML
}
