% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDisMatrix.R
\name{createDisMatrix}
\alias{createDisMatrix}
\title{Dissimilarity matrix}
\usage{
createDisMatrix(ensemble, data)
}
\arguments{
\item{ensemble}{is an ensemble tree object (for the moment ensemble works only with random forest objects)}

\item{data}{a data frame containing the variables in the model. It is the data frame used for ensemble learning.}
}
\value{
A dissimilarity matrix. This is a dissimilarity matrix measuring the discordance between two observations concerning a given classifier of a random forest model.
}
\description{
The function createDisMatrix creates a dissimilarity matrix among observations from an ensemble tree.
}
