% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDisMatrix.R
\name{createDisMatrix}
\alias{createDisMatrix}
\title{Dissimilarity matrix}
\usage{
createDisMatrix(ensemble, data)
}
\arguments{
\item{ensemble}{is an ensemble tree object (for the moment ensemble works only with rf objects)}

\item{data}{is the training dataset used for ensemble learning}
}
\value{
a dissimilarity matrix.
}
\description{
The function crateDisMatrix creates a dissimilarity matrix among observations from an ensemble tree
}
