% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDisMatrix.R
\name{createDisMatrix}
\alias{createDisMatrix}
\title{Dissimilarity matrix}
\usage{
createDisMatrix(ensemble, data, label, parallel = FALSE)
}
\arguments{
\item{ensemble}{is an ensemble tree object (for the moment ensemble works only with random forest objects)}

\item{data}{a data frame containing the variables in the model. It is the data frame used for ensemble learning.}

\item{label}{is a character. It indicates the response label.}

\item{parallel}{to speed up your code. Run all the process in parallel on your laptop using your cores-1}
}
\value{
A dissimilarity matrix. This is a dissimilarity matrix measuring the discordance between two observations concerning a given classifier of a random forest model.
}
\description{
The function createDisMatrix creates a dissimilarity matrix among observations from an ensemble tree.
}
\examples{

## Classification
data(iris)

# Create training and validation set:
data_set_size <- floor(nrow(iris)/2)
indexes <- sample(1:nrow(iris), size = data_set_size)
training <- iris[indexes,]
validation <- iris[-indexes,]
response_training <- training[,5]
response_validation <- validation[,5]

# Perform training:
\dontrun{
require(randomForest)
ensemble <- randomForest(Species ~ ., data=training, importance=TRUE, proximity=TRUE)
D <- createDisMatrix(ensemble, data=data, label = "Species", parallel = TRUE)


# Regression
data("mtcars")

require(randomForest)
ensemble = randomForest(mpg ~ ., data=mtcars, ntree=1000, importance=TRUE, proximity=TRUE)
D = createDisMatrix(ensemble, data=data, label = "mpg", parallel = TRUE)
}
}
