% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2tree.R
\name{e2tree}
\alias{e2tree}
\title{Explainable Ensemble Tree}
\usage{
e2tree(
  formula,
  data,
  D,
  setting = list(impTotal = 0.1, maxDec = 0.01, n = 5, level = 5, tMax = 5),
  method = "classification"
)
}
\arguments{
\item{formula}{is a formula, with a response but no interaction terms.}

\item{data}{is a data frame in which to interpret the variables named in the formula.}

\item{D}{is the dissimilarity matrix.}

\item{setting}{is a list containing the setting parameters for tree building procedure.}

\item{method}{is a character...
Default is \code{setting=list(impTotal=0.1, maxDec=0.01, n=5, level=5, tMax=5)}.}
}
\value{
a e2tree object.

#examples
}
\description{
It creates an explainable tree for Random Forest
}
