% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePredTree.R
\name{ePredTree}
\alias{ePredTree}
\title{Predict responses through an explainable RF}
\usage{
ePredTree(tree, X, target = "1")
}
\arguments{
\item{tree}{is a e2tree object}

\item{X}{is the training data frame with only the predictors}

\item{target}{is the target value of response}
}
\value{
an object.

#examples
}
\description{
It predicts classification tree responses
}
