% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePredTree.R
\name{ePredTree}
\alias{ePredTree}
\title{Predict responses through an explainable RF}
\usage{
ePredTree(fit, data, target = "1")
}
\arguments{
\item{fit}{is a e2tree object}

\item{data}{is a data frame}

\item{target}{is the target value of response}
}
\value{
an object.

#examples
}
\description{
It predicts classification tree responses
}
