% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{get_ensemble_type}
\alias{get_ensemble_type}
\title{Determine Ensemble Type from a Trained Model}
\usage{
get_ensemble_type(ensemble)
}
\arguments{
\item{ensemble}{A trained randomForest or ranger model}
}
\value{
Character string: "classification" or "regression"
}
\description{
Determine Ensemble Type from a Trained Model
}
\keyword{internal}
