% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Roc curve}
\usage{
roc(response, scores, target = "1")
}
\arguments{
\item{response}{is the response variable vector}

\item{scores}{is the probability vector of the prediction}

\item{target}{is the target response class}
}
\value{
an object.

#examples
}
\description{
It calculates roc curve
}
