% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart2Tree.R
\name{rpart2Tree}
\alias{rpart2Tree}
\title{Convert a e2tree into an rpart object}
\usage{
rpart2Tree(tree)
}
\arguments{
\item{tree}{is e2tree object}
}
\value{
an rpart object.

#examples
}
\description{
It converts a e2tree into an rpart object
}
