% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.R
\name{vimp}
\alias{vimp}
\title{Variable Importance}
\usage{
vimp(fit, data)
}
\arguments{
\item{fit}{is a e2tree object}

\item{data}{is a data frame in which to interpret the variables named in the formula.}
}
\value{
a data frame containing variable importance metrics.

#examples
}
\description{
It calculate variable importance of an explainable tree
}
