% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp.R
\name{vimp}
\alias{vimp}
\title{Variable Importance}
\usage{
vimp(tree, response, X, graph = TRUE)
}
\arguments{
\item{tree}{is a e2tree object}

\item{response}{is the vector of responses of the training dataset}

\item{X}{is the training data frame with only the predictors}

\item{graph}{is logical. If TRUE the function will plot a vimp graph.}
}
\value{
a data frame containing variable importance metrics.

#examples
}
\description{
It calculate variable importance of an explainable tree
}
